chore: nx 18 update

+++ features/feature-todo-list/src/lib/todo-list/todo-list-delete.component.ts
@@ -1,28 +1,31 @@
 import {Component, Input} from "@angular/core";
 import {FeatureTodoDelete, FeatureTodoDeletePayload, FeatureTodoDeleteResult} from "@demo/features/feature-todo-common";
 import {filterNill, HasErrorPipe, IsLoadingPipe} from "@demo/utils/utils-data-service";
-import {AsyncPipe, JsonPipe, NgIf} from "@angular/common";
+import { AsyncPipe, JsonPipe } from "@angular/common";
 import {BehaviorSubject, Observable, switchMap} from "rxjs";
 
 @Component({
   selector: 'feature-todo-list-delete',
   template: `
-    <span
-      class="todo-list-delete"
-      *ngIf="params$ | async as params"
-    >
-      <ng-container *ngIf="deleteResult$ | async as deleteResult">
-        <span *ngIf="deleteResult.data | hasError" [title]="deleteResult.data?.error?.message ?? ''"
-              class="warning icon-exclamation"></span>
-        <button
-          class="btn-icon"
-          [disabled]="deleteResult.data | isLoading"
-          (click)="onDeleteButtonClick(params)">
-          <i class="icon-remove"></i>
-        </button>
-      </ng-container>
-    </span>
-  `,
+    @if (params$ | async; as params) {
+      <span
+        class="todo-list-delete"
+        >
+        @if (deleteResult$ | async; as deleteResult) {
+          @if (deleteResult.data | hasError) {
+            <span [title]="deleteResult.data?.error?.message ?? ''"
+            class="warning icon-exclamation"></span>
+          }
+          <button
+            class="btn-icon"
+            [disabled]="deleteResult.data | isLoading"
+            (click)="onDeleteButtonClick(params)">
+            <i class="icon-remove"></i>
+          </button>
+        }
+      </span>
+    }
+    `,
   styles: [`
     .todo-list-delete {
       display: flex;
@@ -34,9 +37,8 @@ import {BehaviorSubject, Observable, switchMap} from "rxjs";
     AsyncPipe,
     HasErrorPipe,
     IsLoadingPipe,
-    NgIf,
     JsonPipe
-  ]
+]
 })
 export class TodoListDeleteComponent {
   readonly params$ = new BehaviorSubject<FeatureTodoDeletePayload | null>(null);
